requirements.txt
Django~=4.2.11
PyJWT~=2.8.0
PyYAML~=6.0.1
asgiref~=3.8.1
attrs~=23.2.0
certifi~=2024.2.2
cffi~=1.16.0
charset-normalizer~=3.3.2
colorama~=0.4.6
cryptography~=3.4.8
djangorestframework~=3.15.1
djangorestframework-simplejwt~=5.3.1
drf-spectacular~=0.27.2
exceptiongroup~=1.2.1
idna~=3.7
inflection~=0.5.1
iniconfig~=2.0.0
jsonschema~=4.22.0
jsonschema-specifications~=2023.12.1
packaging~=24.0
pip~=24.0
pluggy~=1.5.0
psycopg2-binary~=2.9.9
pycparser~=2.22
pytest~=8.2.0
python-decouple~=3.8
python-dotenv~=1.0.1
referencing~=0.35.0
requests~=2.31.0
rpds-py~=0.18.0
setuptools~=69.1.1
sqlparse~=0.5.0
tomli~=2.0.1
typing-extensions~=4.11.0
tzdata~=2024.1
uritemplate~=4.1.1
urllib3~=2.2.1
wheel~=0.42.0
celery~=5.4.0